chore: header rework
