+++ src/pages/about/index.jsx
@@ -5,7 +5,7 @@ import { graphql, useStaticQuery } from "gatsby";
 import Layout from "../../components/layout";
 import PageMetadata from "../../components/page-metadata";
 
-import Header from "../../components/pages/about-us/header";
+import Header from "../../components/header";
 import WhoWeAre from "../../components/pages/about-us/who-we-are";
 import WhereDoWeComeFrom from "../../components/pages/about-us/where-do-we-come-from";
 import WhereAreWeDreamingOfGoingTogether from "../../components/pages/about-us/where-are-we-dreaming-of-going-together";
@@ -35,15 +35,21 @@ const About = () => {
         }
     `);
 
-    console.log(contentfulAboutUsPage);
-
     return (
         <Layout>
             <PageMetadata
                 title={contentfulAboutUsPage.metaTitle.metaTitle}
                 description={contentfulAboutUsPage.metaDescr.metaDescr}
             />
-            <Header />
+            <Header
+                left={
+                    contentfulAboutUsPage.leftHeader.childMarkdownRemark.htmlAst
+                }
+                right={
+                    contentfulAboutUsPage.rightHeader.childMarkdownRemark
+                        .htmlAst
+                }
+            />
             <WhoWeAre />
             <WhereDoWeComeFrom />
             <WhereAreWeDreamingOfGoingTogether />
